requirements.txt
beautifulsoup4==4.12.3
PyYAML==6.0.1
PyYAML==6.0.2
PyYAML==6.0.2
selenium==4.23.1
